+++ R/transform.R
@@ -93,9 +93,8 @@
 #'  plotLDA(cls = 'day')
 #' @export
 
-setGeneric("transformArcSine", function(d) {
-  standardGeneric("transformArcSine")
-})
+setGeneric("transformArcSine", function(d)
+  standardGeneric("transformArcSine"))
 
 #' @rdname transform
 
@@ -109,9 +108,8 @@ setMethod('transformArcSine',signature = 'AnalysisData',
 #' @rdname transform
 #' @export
 
-setGeneric("transformAuto", function(d) {
-  standardGeneric("transformAuto")
-})
+setGeneric("transformAuto", function(d)
+  standardGeneric("transformAuto"))
 
 #' @rdname transform
 
@@ -125,9 +123,8 @@ setMethod('transformAuto',signature = 'AnalysisData',
 #' @rdname transform
 #' @export
 
-setGeneric("transformCenter", function(d) {
-  standardGeneric("transformCenter")
-})
+setGeneric("transformCenter", function(d)
+  standardGeneric("transformCenter"))
 
 #' @rdname transform
 
@@ -141,9 +138,8 @@ setMethod('transformCenter',signature = 'AnalysisData',
 #' @rdname transform
 #' @export
 
-setGeneric("transformLevel", function(d) {
-  standardGeneric("transformLevel")
-})
+setGeneric("transformLevel", function(d)
+  standardGeneric("transformLevel"))
 
 #' @rdname transform
 
@@ -157,9 +153,8 @@ setMethod('transformLevel',signature = 'AnalysisData',
 #' @rdname transform
 #' @export
 
-setGeneric("transformLn", function(d,add = 1) {
-  standardGeneric("transformLn")
-})
+setGeneric("transformLn", function(d,add = 1)
+  standardGeneric("transformLn"))
 
 #' @rdname transform
 
@@ -174,9 +169,8 @@ setMethod('transformLn',signature = 'AnalysisData',
 #' @rdname transform
 #' @export
 
-setGeneric("transformLog10", function(d,add = 1) {
-  standardGeneric("transformLog10")
-})
+setGeneric("transformLog10", function(d,add = 1)
+  standardGeneric("transformLog10"))
 
 #' @rdname transform
 
@@ -191,9 +185,8 @@ setMethod('transformLog10',signature = 'AnalysisData',
 #' @rdname transform
 #' @export
 
-setGeneric("transformPareto", function(d) {
-  standardGeneric("transformPareto")
-})
+setGeneric("transformPareto", function(d)
+  standardGeneric("transformPareto"))
 
 #' @rdname transform
 
@@ -210,9 +203,8 @@ setMethod('transformPareto',signature = 'AnalysisData',
 #' @rdname transform
 #' @export
 
-setGeneric("transformRange", function(d) {
-  standardGeneric("transformRange")
-})
+setGeneric("transformRange", function(d)
+  standardGeneric("transformRange"))
 
 #' @rdname transform
 
@@ -230,9 +222,8 @@ setMethod('transformRange',signature = 'AnalysisData',
 #' @rdname transform
 #' @export
 
-setGeneric("transformSQRT", function(d) {
-  standardGeneric("transformSQRT")
-})
+setGeneric("transformSQRT", function(d)
+  standardGeneric("transformSQRT"))
 
 #' @rdname transform
 
@@ -246,9 +237,8 @@ setMethod('transformSQRT',signature = 'AnalysisData',
 #' @rdname transform
 #' @export
 
-setGeneric("transformTICnorm", function(d) {
-  standardGeneric("transformTICnorm")
-})
+setGeneric("transformTICnorm", function(d)
+  standardGeneric("transformTICnorm"))
 
 #' @rdname transform
 
@@ -267,9 +257,8 @@ setMethod('transformTICnorm',signature = 'AnalysisData',
 #' @rdname transform
 #' @export
 
-setGeneric("transformVast", function(d) {
-  standardGeneric("transformVast")
-})
+setGeneric("transformVast", function(d)
+  standardGeneric("transformVast"))
 
 #' @rdname transform